src/sonar_bertalign/embeddings.py
import numpy as np
import torch
# Attempt to import SONAR, with a placeholder if not found initially
try:
    from sonar.models import load_text_encoder
    SONAR_AVAILABLE = True
except ImportError:
    print("Warning: SONAR library not found. SonarEmbedder will not be fully functional.")
    # Define a placeholder if SONAR is not available to avoid runtime errors on class definition
    def load_text_encoder(model_name_or_path, device):
        print(f"Placeholder: Would load SONAR model {model_name_or_path} on {device}")
        # Return a dummy object that mimics the expected encoder structure if possible
        # This is highly dependent on SONAR's actual API
        class DummyEncoder:
            def __call__(self, *args, **kwargs):
                raise NotImplementedError("SONAR not installed, cannot encode.")
            def to(self, device):
                print(f"Placeholder: DummyEncoder moved to {device}")
                return self
        return DummyEncoder()
    SONAR_AVAILABLE = False

class SonarEmbedder:
    """
    A class to generate sentence embeddings using Meta's SONAR models.
    """
    def __init__(self, model_name_or_path: str = "sonar_text_encoder_eng_rus", device: str = 'cpu'):
        """
        Initializes the SonarEmbedder.

        Args:
            model_name_or_path (str): The name or path to the SONAR text encoder model.
                                      Defaults to a hypothetical English-Russian model.
                                      Replace with an actual available SONAR model identifier.
            device (str): The device to load the model on ('cpu' or 'cuda').
        """
        self.device = device
        if SONAR_AVAILABLE:
            try:
                self.model = load_text_encoder(model_name_or_path, device=self.device)
                print(f"SONAR model '{model_name_or_path}' loaded successfully on {self.device}.")
            except Exception as e:
                print(f"Error loading SONAR model '{model_name_or_path}': {e}")
                print("Falling back to placeholder functionality for SonarEmbedder.")
                self.model = load_text_encoder(model_name_or_path, device=self.device) # This will use the dummy
        else:
            self.model = load_text_encoder(model_name_or_path, device=self.device) # Uses the dummy

    def encode(self, sentences: list[str], batch_size: int = 32) -> np.ndarray:
        """
        Encodes a list of sentences into embeddings.

        Args:
            sentences (list[str]): A list of sentences to encode.
            batch_size (int): The batch size for processing.

        Returns:
            np.ndarray: A NumPy array containing the sentence embeddings.
                        Returns an empty array if SONAR is not available or encoding fails.
        """
        if not SONAR_AVAILABLE or not hasattr(self.model, 'encode_text_batch'): # Check for actual encode method
            print("SONAR model not available or does not have 'encode_text_batch' method. Cannot generate embeddings.")
            # Check actual SONAR API for method name
            # For fairseq2 SONAR text encoder, it's likely `model.encode_text_batch(sentences)`
            # or similar after tokenization if not handled internally by `encode_text_batch`.
            # The SONAR API documentation should be consulted for the exact method.
            # This is a placeholder for the actual encoding call.
            # Depending on SONAR's API, sentences might need pre-processing (tokenization)
            # before being passed to the model.
            # The `sonar.models.load_text_encoder` should provide an object that can directly process text.
            return np.array([])

        all_embeddings = []
        for i in range(0, len(sentences), batch_size):
            batch = sentences[i:i+batch_size]
            try:
                # The actual call to SONAR's encoding function needs to be verified.
                # Example: embeddings_tensor = self.model.encode_text_batch(batch) # This is a guess
                # Or it might be:
                # tokenized_batch = self.model.tokenize(batch) # Hypothetical tokenization step
                # embeddings_tensor = self.model(tokenized_batch) # Model forward pass
                # This placeholder assumes self.model directly handles sentence list.
                # Replace with the correct SONAR API call.
                
                # Placeholder: Simulate encoding process
                print(f"Placeholder: Encoding batch of {len(batch)} sentences.")
                # Simulate tensor output then conversion to numpy
                # The shape would be (len(batch), embedding_dim)
                # Embedding dimension for SONAR is typically 1024.
                dummy_embeddings_tensor = torch.randn(len(batch), 1024, device=self.device)
                
                # This part assumes the model returns PyTorch tensors
                if self.device == 'cuda':
                    all_embeddings.append(dummy_embeddings_tensor.cpu().numpy())
                else:
                    all_embeddings.append(dummy_embeddings_tensor.numpy())

            except NotImplementedError:
                 print("SONAR actual encoding method not implemented in this placeholder.")
                 return np.array([]) # Stop if the dummy method is hit
            except Exception as e:
                print(f"Error encoding batch with SONAR: {e}")
                # Optionally, continue to next batch or return empty
                return np.array([]) # Stop on error for now

        if not all_embeddings:
            return np.array([])
            
        return np.vstack(all_embeddings)

if __name__ == '__main__':
    # This basic test will run if the file is executed directly.
    # It uses the placeholder functionality if SONAR is not installed.

    print("\n--- Testing SonarEmbedder ---")
    # Example: use a real model name if SONAR is installed, e.g., "sonar_text_encoder_eng_rus"
    # For testing, we use a placeholder name.
    embedder = SonarEmbedder(model_name_or_path="sonar_text_LID", device='cpu')

    sample_sentences_en = [
        "Hello world!",
        "This is a test sentence."
    ]
    sample_sentences_fr = [
        "Bonjour le monde!",
        "Ceci est une phrase de test."
    ]

    print("\nEncoding English sentences...")
    embeddings_en = embedder.encode(sample_sentences_en)
    if embeddings_en.size > 0:
        print(f"Generated English embeddings of shape: {embeddings_en.shape}")
        # print(embeddings_en)
    else:
        print("Failed to generate English embeddings (SONAR might not be installed or model issue).")

    print("\nEncoding French sentences...")
    embeddings_fr = embedder.encode(sample_sentences_fr)
    if embeddings_fr.size > 0:
        print(f"Generated French embeddings of shape: {embeddings_fr.shape}")
        # print(embeddings_fr)
    else:
        print("Failed to generate French embeddings (SONAR might not be installed or model issue).")

    # Example of how to use with GPU if available and SONAR installed
    # if torch.cuda.is_available() and SONAR_AVAILABLE:
    #     print("\n--- Testing SonarEmbedder on CUDA ---")
    #     embedder_gpu = SonarEmbedder(model_name_or_path="sonar_text_LID", device='cuda')
    #     print("\nEncoding English sentences on GPU...")
    #     embeddings_en_gpu = embedder_gpu.encode(sample_sentences_en)
    #     if embeddings_en_gpu.size > 0:
    #         print(f"Generated English embeddings on GPU of shape: {embeddings_en_gpu.shape}")
    #     else:
    #         print("Failed to generate English embeddings on GPU.")
    # else:
    #     if not torch.cuda.is_available():
    #         print("\nCUDA not available, skipping GPU test.")
    #     if not SONAR_AVAILABLE:
    #         print("\nSONAR not available, skipping GPU test.") 